Add more packages
- Abstract enum commands into seperate action class files
- Move DukeDB and DukeException to util package
- Add task package
