+++ src/main/java/duke/Main.java
@@ -1,5 +1,7 @@
 package duke;
 
+import duke.util.DukeDB;
+
 public class Main {
     public static void main(String[] args) {
         DukeDB db = new DukeDB("./data/dukeStore.txt");
