+++ src/main/java/duke/command/ListCommand.java
@@ -1,0 +1,18 @@
+package duke.command;
+
+import duke.util.DukeDB;
+import duke.util.DukeException;
+import duke.util.DukeConfig;
+import duke.util.DukeTaskList;
+import duke.util.Ui;
+
+import java.util.Map;
+
+public class ListCommand implements DukeActions {
+    @Override
+    public boolean run(Map<String, String> map, DukeTaskList list, DukeDB database, DukeConfig config, Ui ui)
+            throws DukeException {
+        ui.printTaskList(list);
+        return true;
+    }
+}
