+++ src/main/java/duke/util/DukeDB.java
@@ -1,6 +1,10 @@
-package duke;
+package duke.util;
 
-import duke.util.Parser;
+import duke.Duke;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.Todo;
 
 import java.io.File;
 import java.io.FileNotFoundException;
