+++ src/main/java/duke/util/DukeTaskList.java
@@ -1,6 +1,6 @@
 package duke.util;
 
-import duke.Task;
+import duke.task.Task;
 
 import java.util.ArrayList;
 import java.util.Optional;
