+++ src/main/java/duke/util/Parser.java
@@ -1,7 +1,5 @@
 package duke.util;
 
-import duke.DukeException;
-
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
